Move FPN to ops
