+++ torchvision/ops/__init__.py
@@ -2,9 +2,10 @@ from .boxes import nms, box_iou
 from .roi_align import roi_align, RoIAlign
 from .roi_pool import roi_pool, RoIPool
 from .poolers import MultiScaleRoIAlign
+from .feature_pyramid_network import FeaturePyramidNetwork
 
 
 __all__ = [
     'nms', 'roi_align', 'RoIAlign', 'roi_pool', 'RoIPool',
-    'MultiScaleRoIAlign'
+    'MultiScaleRoIAlign', 'FeaturePyramidNetwork'
 ]
